login mechanism cleanup
