+++ index.html
@@ -3,15 +3,15 @@
   <head>
       <title>Tandem SDK Test Bed</title>
       <!--<script src="node_modules/@adsk/lmv-double-trouble/build/viewer3D.js"></script>-->
-      <!--<script src="https://tandem.autodesk.com/viewer/viewer3D.js"></script>-->
-      <script src="https://tandem-stg.autodesk.com/viewer/viewer3D.js"></script>
+      <script src="https://tandem.autodesk.com/viewer/viewer3D.js"></script>
+      <!--<script src="https://tandem-stg.autodesk.com/viewer/viewer3D.js"></script>-->
       <script src="./3rdParty/jquery-3.6.0.min.js"></script>
       <script src="./3rdParty/bootstrap-4.0.0/dist/js/bootstrap.bundle.js"></script>
 
       <!-- CSS only -->
       <!--<link href="node_modules/@adsk/lmv-double-trouble/build/style.min.css" rel="stylesheet" type="text/css">-->
-      <!--<link href="https://tandem.autodesk.com/viewer/style.min.css" rel="stylesheet" type="text/css">-->
-      <link href="https://tandem-stg.autodesk.com/viewer/style.min.css" rel="stylesheet" type="text/css">
+      <link href="https://tandem.autodesk.com/viewer/style.min.css" rel="stylesheet" type="text/css">
+      <!--<link href="https://tandem-stg.autodesk.com/viewer/style.min.css" rel="stylesheet" type="text/css">-->
       <link href="./3rdParty/bootstrap-4.0.0/dist/css/bootstrap.css" rel="stylesheet" type="text/css">
   </head>
   <body>
