--- package.json
@@ -1,14 +1,0 @@
-{
-  "name": "tandem-sample-emb-Viewer",
-  "version": "0.1.0",
-  "description": "Demonstration of Embedded Viewer API",
-  "main": "index.js",
-  "scripts": {
-    "start": "static -p 8080"
-  },
-  "author": "james.awe@autodesk.com",
-  "license": "Unlicensed",
-  "dependencies": {
-    "node-static": "^0.7.11"
-  }
-}
